refactor: move chainparams into coin's folder
