+++ basicswap/interface/passthrough_btc.py
@@ -4,7 +4,7 @@
 # Distributed under the MIT software license, see the accompanying
 # file LICENSE or http://www.opensource.org/licenses/mit-license.php.
 
-from .btc import BTCInterface
+from basicswap.interface.btc.btc import BTCInterface
 from basicswap.contrib.test_framework.messages import CTxOut
